Separated scroll rect enabled functionality from AddLayerPanel and added it to LayerPanel
