+++ Assets/Prefabs/Layers/UI/AddLayerPanel.prefab
@@ -469,8 +469,8 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 1
         m_CallState: 2
-      - m_Target: {fileID: 67748899823263108}
-        m_TargetAssemblyTypeName: Netherlands3D.Twin.AddLayerPanel, Assembly-CSharp
+      - m_Target: {fileID: 1372417627372885277}
+        m_TargetAssemblyTypeName: Netherlands3D.Twin.AdvancedScrollView, Assembly-CSharp
         m_MethodName: ResetScrollActive
         m_Mode: 1
         m_Arguments:
@@ -794,6 +794,7 @@ GameObject:
   - component: {fileID: 1890818014237057293}
   - component: {fileID: 675677739205232157}
   - component: {fileID: 67748899823263108}
+  - component: {fileID: 1372417627372885277}
   m_Layer: 5
   m_Name: AddLayerPanel
   m_TagString: Untagged
@@ -913,6 +914,18 @@ MonoBehaviour:
   OnRectTransformSizeChanged:
     m_PersistentCalls:
       m_Calls: []
+--- !u!114 &1372417627372885277
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 740642816338865571}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3790799467c1e48d2886a5d54fb7a9b6, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &864163037451147846
 GameObject:
   m_ObjectHideFlags: 0
@@ -7305,8 +7318,8 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
-      - m_Target: {fileID: 67748899823263108}
-        m_TargetAssemblyTypeName: Netherlands3D.Twin.AddLayerPanel, Assembly-CSharp
+      - m_Target: {fileID: 1372417627372885277}
+        m_TargetAssemblyTypeName: Netherlands3D.Twin.AdvancedScrollView, Assembly-CSharp
         m_MethodName: ResetScrollActive
         m_Mode: 1
         m_Arguments:
@@ -8496,8 +8509,8 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 1
         m_CallState: 2
-      - m_Target: {fileID: 67748899823263108}
-        m_TargetAssemblyTypeName: Netherlands3D.Twin.AddLayerPanel, Assembly-CSharp
+      - m_Target: {fileID: 1372417627372885277}
+        m_TargetAssemblyTypeName: Netherlands3D.Twin.AdvancedScrollView, Assembly-CSharp
         m_MethodName: ResetScrollActive
         m_Mode: 1
         m_Arguments:
