+++ Assets/Scripts/UI/AdvancedScrollView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3790799467c1e48d2886a5d54fb7a9b6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
